+++ src/Core/Auras/AuraTtsSounds.lua
@@ -159,7 +159,9 @@ addon.Core.AuraTtsSounds = {
 		[208086] = "ColossusSmash.ogg", -- Colossus Smash
 		[77606] = "DarkSimulacrum.ogg", -- Dark Simulacrum
 		[360194] = "Deathmark.ogg", -- Deathmark
+		[236273] = "Duel.ogg", -- Duel
 		[274837] = "FeralFrenzy.ogg", -- Feral Frenzy
 		[385627] = "Kingsbane.ogg", -- Kingsbane
+		[198819] = "SharpenBlade.ogg", -- Sharpen Blade
 	},
 }